requirements.txt
Django>=4.2
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.25.0
reportlab>=4.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
openpyxl>=3.1.0
cerebras_cloud_sdk
